extract server-related stuff
