+++ greetsgraf.go
@@ -2,17 +2,10 @@ package main
 
 import (
 	"compress/gzip"
-	"context"
 	"encoding/json"
 	"flag"
-	"github.com/go-chi/chi/v5"
-	"github.com/go-chi/chi/v5/middleware"
-	"gorm.io/gorm"
 	"log"
-	"net/http"
 	"os"
-	"strconv"
-	"time"
 )
 
 func readJsonGz(filename string) (map[string]interface{}, error) {
@@ -36,313 +29,6 @@ func readJsonGz(filename string) (map[string]interface{}, error) {
 	}
 
 	return value, err
-}
-
-func respondErrJson(w http.ResponseWriter, status int, err error) {
-	response, jerr := json.Marshal(struct{ Error string }{Error: err.Error()})
-	if jerr != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte(err.Error()))
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	w.Write([]byte(response))
-}
-
-func respondJson(w http.ResponseWriter, status int, payload interface{}) {
-	response, err := json.Marshal(payload)
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	w.Write([]byte(response))
-}
-
-func (c *Database) groupsFind(w http.ResponseWriter, r *http.Request) {
-	query := r.URL.Query()
-	name := query.Get("name")
-	if name == "" {
-		respondJson(w, http.StatusOK, []int{})
-		return
-	}
-
-	groups, err := c.FindGroups(name)
-	if err != nil {
-		// TODO proper error status
-		respondJson(w, http.StatusInternalServerError, []int{})
-		return
-	}
-
-	respondJson(w, http.StatusOK, &groups)
-}
-
-func (c *Database) findProd(w http.ResponseWriter, r *http.Request) {
-	query := r.URL.Query()
-	name := query.Get("name")
-	if name == "" {
-		respondJson(w, http.StatusOK, []int{})
-		return
-	}
-
-	prods, err := c.FindProds(name)
-	if err != nil {
-		// TODO proper error status
-		respondJson(w, http.StatusInternalServerError, []int{})
-		return
-	}
-
-	respondJson(w, http.StatusOK, &prods)
-}
-
-func (c *Database) prodGet(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	pid := ctx.Value("prod_id")
-
-	prod, err := c.GetProd(pid)
-	if err != nil {
-		// TODO proper error status
-		respondJson(w, http.StatusInternalServerError, struct{}{})
-		return
-	}
-
-	// TODO maybe it's better done through a custom marshaller ...
-	type ResponseGroup struct {
-		ID             uint
-		Name           string
-		Disambiguation string
-	}
-
-	type ResponseGreet struct {
-		ID    uint
-		Group ResponseGroup
-		Note  string
-	}
-
-	response_prod := struct {
-		ID         uint
-		Name       string
-		Year       int
-		Month      int
-		Day        int
-		Video      string
-		Rank       int
-		VoteUp     int
-		VotePig    int
-		VoteDown   int
-		Demozoo    int
-		Screenshot string
-		Groups     []ResponseGroup
-		Greets     []ResponseGreet
-	}{
-		ID:         prod.ID,
-		Name:       prod.Name,
-		Year:       prod.Year,
-		Month:      prod.Month,
-		Day:        prod.Day,
-		Video:      prod.Video,
-		Rank:       prod.Rank,
-		VoteUp:     prod.VoteUp,
-		VotePig:    prod.VotePig,
-		VoteDown:   prod.VoteDown,
-		Demozoo:    prod.Demozoo,
-		Screenshot: prod.Screenshot,
-	}
-
-	for i, _ := range prod.Groups {
-		group := &prod.Groups[i]
-		response_prod.Groups = append(response_prod.Groups, ResponseGroup{
-			ID:             group.ID,
-			Name:           group.Name,
-			Disambiguation: group.Disambiguation,
-		})
-	}
-
-	for i, _ := range prod.Greets {
-		greet := &prod.Greets[i]
-		group, err := c.GetGroup(greet.GreeteeID)
-		if err != nil {
-			log.Println(err)
-			continue
-		}
-		response_prod.Greets = append(response_prod.Greets, ResponseGreet{
-			ID:   greet.ID,
-			Note: greet.Reference,
-			Group: ResponseGroup{
-				ID:             group.ID,
-				Name:           group.Name,
-				Disambiguation: group.Disambiguation,
-			},
-		})
-	}
-
-	respondJson(w, http.StatusOK, response_prod)
-}
-
-func (c *Database) prodGetGreets(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	prod_id := ctx.Value("prod_id")
-
-	greets, err := c.GetProdGreets(prod_id)
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	respondJson(w, http.StatusOK, greets)
-}
-
-func (c *Database) groupGetGreeted(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	group_id := ctx.Value("group_id")
-
-	greets, err := c.GetGroupGreets(group_id)
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	respondJson(w, http.StatusOK, &greets)
-}
-
-func (c *Database) greetsCreate(w http.ResponseWriter, r *http.Request) {
-	var body struct {
-		ProdId  uint
-		GroupId uint
-		Note    string
-	}
-	err := json.NewDecoder(r.Body).Decode(&body)
-	if err != nil {
-		respondErrJson(w, http.StatusBadRequest, err)
-		return
-	}
-
-	id, err := c.Greet(body.ProdId, body.GroupId, body.Note)
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	respondJson(w, http.StatusOK, struct{ ID uint }{id})
-}
-
-func (c *Database) greetsDelete(w http.ResponseWriter, r *http.Request) {
-	greet_id, err := strconv.Atoi(chi.URLParam(r, "id"))
-	if err != nil {
-		respondErrJson(w, http.StatusBadRequest, err)
-		return
-	}
-
-	removed, err := c.DeleteGreet(uint(greet_id))
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	if !removed {
-		respondJson(w, http.StatusNotFound, struct{}{})
-	}
-
-	respondJson(w, http.StatusOK, struct{}{})
-}
-
-func (c *Database) getStats(w http.ResponseWriter, r *http.Request) {
-	stats := c.GetStats()
-	respondJson(w, http.StatusOK, stats)
-}
-
-func (c *Database) groupsGreeted(w http.ResponseWriter, r *http.Request) {
-	query := r.URL.Query()
-	limit, _ := strconv.Atoi(query.Get("limit"))
-
-	results, err := c.GetMostGreetedGroups(limit)
-
-	if err != nil {
-		respondErrJson(w, http.StatusInternalServerError, err)
-		return
-	}
-
-	respondJson(w, http.StatusOK, results)
-}
-
-func ProdContext(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		prod_id, err := strconv.Atoi(chi.URLParam(r, "id"))
-		if err != nil {
-			respondErrJson(w, http.StatusBadRequest, err)
-			return
-		}
-
-		ctx := context.WithValue(r.Context(), "prod_id", prod_id)
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
-func GroupContext(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		group_id, err := strconv.Atoi(chi.URLParam(r, "id"))
-		if err != nil {
-			respondErrJson(w, http.StatusBadRequest, err)
-			return
-		}
-
-		ctx := context.WithValue(r.Context(), "group_id", group_id)
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
-
-func listen(db *gorm.DB, listen string, serve_static string) {
-	ctx := Database{db}
-
-	r := chi.NewRouter()
-
-	r.Use(middleware.Logger)
-	r.Use(middleware.Timeout(10 * time.Second))
-
-	r.Route("/v1", func(r chi.Router) {
-		r.Get("/stats", ctx.getStats)
-
-		r.Route("/groups", func(r chi.Router) {
-			r.Get("/search", ctx.groupsFind)
-			r.Get("/greeted", ctx.groupsGreeted)
-			r.Route("/{id}", func(r chi.Router) {
-				r.Use(GroupContext)
-				//r.Get("/", ctx.groupGet)
-				r.Get("/greets", ctx.groupGetGreeted)
-			})
-		})
-		r.Route("/prods", func(r chi.Router) {
-			r.Get("/search", ctx.findProd)
-			r.Route("/{id}", func(r chi.Router) {
-				r.Use(ProdContext)
-				r.Get("/", ctx.prodGet)
-				r.Get("/greets", ctx.prodGetGreets)
-			})
-		})
-
-		r.Route("/greets", func(r chi.Router) {
-			r.Post("/", ctx.greetsCreate)
-			r.Route("/{id}", func(r chi.Router) {
-				//r.Get("", ctx.greetsGet)
-				//r.Patch("", ctx.greetsUpdate)
-				r.Delete("/", ctx.greetsDelete)
-			})
-		})
-	})
-
-	if serve_static != "" {
-		fs := http.FileServer(http.Dir(serve_static))
-		r.Get("/*", func(w http.ResponseWriter, r *http.Request) {
-			fs.ServeHTTP(w, r)
-		})
-	}
-
-	log.Printf("Listening on %+v", listen)
-	log.Fatal(http.ListenAndServe(listen, r))
 }
 
 type Args struct {
@@ -393,6 +79,6 @@ func main() {
 	}
 
 	if args.serve {
-		listen(db, args.listen, args.static)
+		listen(Database{db}, args.listen, args.static)
 	}
 }
